notification fixes
