+++ pluto/lib/src/main/java/com/pluto/core/notification/NotificationManager.kt
@@ -1,0 +1,21 @@
+package com.pluto.core.notification
+
+import android.app.Application
+import androidx.lifecycle.MutableLiveData
+import com.pluto.utilities.AppState
+
+@SuppressWarnings("UseDataClass")
+internal class NotificationManager(application: Application, state: MutableLiveData<AppState>) {
+
+    private val debugNotification = DebugNotification(application.applicationContext)
+
+    init {
+        state.observeForever {
+            if (it is AppState.Foreground) {
+                debugNotification.add()
+            } else {
+                debugNotification.remove()
+            }
+        }
+    }
+}
